update btrfsprogs
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@31112 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/btrfsprogs/tests/mkfs-tests/006-partitioned-loopdev/test.sh
@@ -3,8 +3,12 @@
 
 source $TOP/tests/common
 
+if ! losetup --help | grep -q 'partscan'; then
+	_not_run "losetup --partscan not available"
+	exit 0
+fi
+
 check_prereq mkfs.btrfs
-check_prereq btrfs-show-super
 
 setup_root_helper
 
@@ -19,7 +23,7 @@ base=$(basename $loopdev)
 # expect partitions named like loop0p1 etc
 for looppart in $(ls /dev/$base?*); do
 	run_check $SUDO_HELPER $TOP/mkfs.btrfs -f $looppart
-	run_check $TOP/btrfs-show-super $looppart
+	run_check $SUDO_HELPER $TOP/btrfs inspect-internal dump-super $looppart
 done
 
 # cleanup
